refactor: feed footer and feed refresh control
